Refactor: Split `object.carve` operator as `carve_box` and `carve_polyline`
For circle object third operator will be created, which will not require modal support.
